backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pymongo==4.3.3
pandas==2.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
matplotlib==3.7.1
numpy==1.24.3
openpyxl==3.1.2
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
email-validator==1.3.1
pydantic==1.10.7
